requirements.txt
boto3==1.17.78
botocore==1.20.78
click==8.0.0
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.2
Flask==2.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.0
Jinja2==3.0.0
jmespath==0.10.0
joblib==1.0.1
MarkupSafe==2.0.0
numpy==1.20.3
pandas==1.2.4
psycopg2==2.8.6
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
SQLAlchemy==1.4.15
threadpoolctl==2.1.0
urllib3==1.26.4
Werkzeug==2.0.0
WTForms==2.3.3
